+++ momoka-node/src/proofs/publications/post/da-post-verifier-v2.ts
@@ -5,7 +5,7 @@ import { LogFunctionType } from '../../../common/logger';
 import { failure, PromiseResult, success } from '../../../data-availability-models/da-result';
 import { BigNumber } from 'ethers';
 import { MomokaValidatorError } from '../../../data-availability-models/validator-errors';
-import { DAActionTypes } from '../../../data-availability-models/data-availability-action-types';
+import { MomokaActionTypes } from '../../../data-availability-models/data-availability-action-types';
 import {
   blockHashExists,
   EMPTY_BYTE,
@@ -27,11 +27,13 @@ export type DAPostPublicationV2 = DAStructurePublication<
 export const isDAPostPublicationV2 = (
   daPublication: DAStructurePublication
 ): daPublication is DAPostPublicationV2 => {
-  return daPublication.type === DAActionTypes.POST_CREATED && 'postParams' in daPublication.event;
+  return (
+    daPublication.type === MomokaActionTypes.POST_CREATED && 'postParams' in daPublication.event
+  );
 };
 
 export class DAPostVerifierV2 extends DAPublicationVerifierV2 {
-  public readonly type = DAActionTypes.POST_CREATED;
+  public readonly type = MomokaActionTypes.POST_CREATED;
 
   constructor(
     public readonly daPublication: DAPostPublicationV2,