+++ momoka-node/CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased - fill in any changes you do here!
 
 - feat: add support for lens v2 publications
+- fix: export MomokaProvider and MomokaActionTypes
 
 ## Features